# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Social Documentation",
        "description": "Social description",
        "contact": {
            "email": "admin@admin.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/profile/{id}/relationships": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Gets user's relationships",
                "description": "Gets user's relationships",
                "operationId": "getUserRelationships",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userId": {
                                            "type": "string"
                                        },
                                        "friendId": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "dateRequested": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "dateAccepted": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "dateRejected": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "dateBlocked": {
                                            "type": "string",
                                            "format": "date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/profile/{id}/relationships/{friend}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Gets the relationship",
                "description": "Gets the relationship",
                "operationId": "getRelationship",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "friendId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userId": {
                                            "type": "string"
                                        },
                                        "friendId": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "dateRequested": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "dateAccepted": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "dateRejected": {
                                            "type": "string",
                                            "format": "date"
                                        },
                                        "dateBlocked": {
                                            "type": "string",
                                            "format": "date"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Request user as friend",
                "description": "Request user as friend",
                "operationId": "requestFriend",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "friendId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Update user's friend status - accept/reject/block",
                "description": "Update user's friend status - accept/reject/block",
                "operationId": "updateFriend",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "friendId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Unblock user's friend - removing their relationship",
                "description": "Unblock user's friend - removing their relationship",
                "operationId": "unblockUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "friendId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
        "/profile/{id}/reports": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Gets the reports that were submitted by the user",
                "description": "Gets the reports that were submitted by the user",
                "operationId": "getUserReports",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userId": {
                                            "type": "string"
                                        },
                                        "info": {
                                            "type": "string"
                                        },
                                        "dateCreated": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Profile"
                ],
                "summary": "Report a user",
                "description": "Report a user",
                "operationId": "reportUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}